piksi_system_daemon: Move sigchld handling to async-child
Protect access to `async_children` by masking SIGCHLD.
